Renamed apply confirm
